OcrFunctions/requirements.txt
adal==1.2.0
asn1crypto==0.24.0
astroid==2.1.0
atomicwrites==1.2.1
attrs==18.2.0
azure-common==1.1.16
azure-eventgrid==1.2.0
azure-functions==1.0.0a5
azure-functions-worker==1.0.0a6
azure-nspkg==3.0.2
azure-storage-blob==1.4.0
azure-storage-common==1.4.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
colorama==0.4.1
coverage==4.5.2
cryptography==2.4.2
grpcio==1.14.2
grpcio-tools==1.14.2
idna==2.8
isodate==0.6.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
mockito==1.1.1
more-itertools==4.3.0
msrest==0.6.2
msrestazure==0.5.1
oauthlib==2.1.0
Pillow==5.3.0
pluggy==0.8.0
protobuf==3.6.1
py==1.7.0
pycparser==2.19
PyJWT==1.7.1
pylint==2.2.2
pytest==4.0.2
pytest-cov==2.6.0
python-dateutil==2.7.5
requests==2.21.0
requests-oauthlib==1.0.0
six==1.12.0
typed-ast==1.1.0
urllib3==1.24.1
wrapt==1.10.11
